refactor: separate Project into pure domain, entity and add MapStruct mapper

+++ src/main/java/com/mx/tecdesoftware/knowtion/models/Project.java
@@ -8,10 +8,11 @@ import java.util.List;
 @Entity
 @Table(name = "projects")
 public class Project {
+    
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String titulo;
@@ -42,8 +43,8 @@ public class Project {
     }
 
     // --- Getters y Setters ---
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
     public String getTitulo() { return titulo; }
     public void setTitulo(String titulo) { this.titulo = titulo; }
     public String getDescripcion() { return descripcion; }
